requirements.txt
streamlit
pandas
numpy
faiss-cpu
requests
tqdm
